Fix views
