+++ rocky/tests/objects/test_objects_tree.py
@@ -34,7 +34,7 @@ def test_ooi_tree(rf, client_member, mock_organization_view_octopoes):
     response = OOITreeView.as_view()(request, organization_code=client_member.organization.code)
 
     assert response.status_code == 200
-    assert mock_organization_view_octopoes().get_tree.call_count == 1
+    assert mock_organization_view_octopoes().get_tree.call_count == 2
     assertContains(response, "testnetwork")
     assertContains(response, "KAT-000")
